requirements.txt
altair==5.5.0
aiohttp
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
fonttools==4.58.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
groq==0.25.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx[http2]
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.40.0
numpy==2.2.6
oauthlib==3.2.2
openai==1.79.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.31.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
audioop-lts==0.2.1
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.25.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
watchdog==6.0.0
youtube-transcript-api>=1.1.0
yt-dlp>=2025.6.25
